requirements.txt
# Core
flask>=2.3.0
flask-cors>=4.0.0
pymongo>=4.3.3
requests>=2.28.1
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing
pandas>=1.3.3
numpy>=1.21.2

# Matching / ML (legacy)
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
scikit-learn>=0.24.2
rapidfuzz>=3.0.0

# Scraping - Standard
bs4>=0.0.1
beautifulsoup4>=4.12.0
httpx>=0.27.0
lxml>=5.1.0
tenacity>=8.2.3

# Scraping - Hardcore (Playwright)
playwright>=1.40.0
pytest-playwright>=0.4.0

# Async
aiofiles>=23.0.0

# Utils
matplotlib>=3.4.3
joblib>=1.3.0

